+++ now_lms/db/tools.py
@@ -1,3 +1,17 @@
+# Copyright 2025 BMO Soluciones, S.A.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 # Copyright 2022 - 2024 BMO Soluciones, S.A.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -995,6 +1009,14 @@ def is_blog_enabled():
     """Check if blog are enabled in navigation."""
     if config := _get_global_config():
         return config.enable_blog
+    return False
+
+
+@cache.cached(timeout=300, key_prefix="nav_contact_enabled")
+def is_contact_enabled():
+    """Check if contact page is enabled in navigation."""
+    if config := _get_global_config():
+        return config.enable_contact
     return False