+++ now_lms/forms/__init__.py
@@ -11,6 +11,20 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
+# Copyright 2025 BMO Soluciones, S.A.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 #
 """Definición de formularios."""
 
@@ -392,6 +406,7 @@ class ConfigForm(FlaskForm):
     enable_masterclass = BooleanField(_("Habilitar Master Class"), default=False, validators=[])
     enable_resources = BooleanField(_("Habilitar Recursos descargables"), default=False, validators=[])
     enable_blog = BooleanField(_("Habilitar Blog"), default=False, validators=[])
+    enable_contact = BooleanField(_("Habilitar Contacto"), default=False, validators=[])
 
     # Blog display configuration
     show_latest_blog_posts_on_home = BooleanField(